restore_points/restore_20251002_104500/requirements.txt
opencv-python
numpy
Pillow
pytesseract
pdf2image
PyYAML
rich
requests
